Fix math operands and format a file

+++ CoreOpenGLDisplay/src/au/gov/asd/tac/constellation/visual/opengl/utilities/glyphs/GlyphManagerBI.java
@@ -553,12 +553,12 @@ public final class GlyphManagerBI implements GlyphManager {
 
     @Override
     public float getWidthScalingFactor() {
-        return textureBuffer.width / maxFontHeight;
+        return (float) textureBuffer.width / maxFontHeight;
     }
 
     @Override
     public float getHeightScalingFactor() {
-        return textureBuffer.height / maxFontHeight;
+        return (float) textureBuffer.height / maxFontHeight;
     }
 
     BufferedImage getTextureBuffer() {
